--- a/changelog.d/5038.bugfix
+++ b/changelog.d/5038.bugfix
@@ -1,0 +1,1 @@
+Fixing missing/intermittent notifications on the google play variant when wifi is enabled